% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bear_matrix.R
\docType{data}
\name{bear_matrix}
\alias{bear_matrix}
\title{Extinction probabilites for several communities of bears.}
\format{A taxa by community matrix with 13 bear species for rows. The names match the tip labels for \code{\link{bear_tree}}, an included, dated phylogeny. The columns represent 3 different modern communities of bears: Global, Eurasia, and North America.}
\usage{
bear_matrix
}
\description{
Example extinction (i.e. non occurence) probabilities for late Quaternary bear species for the whole world, Eurasia, and North America.
}
\keyword{datasets}
