% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bear_probs.R
\docType{data}
\name{bear_probs}
\alias{bear_probs}
\title{Probabilites bears are present.}
\format{A named vector with 13 bear species. The names match the tip labels for \code{\link{bear_tree}}, an included dated phylogeny.}
\usage{
bear_probs
}
\description{
Example probabilities of presence (1 - extinction probabilities) for late Quaternary bear species.
}
\keyword{datasets}
