% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bear_tree.R
\docType{data}
\name{bear_tree}
\alias{bear_tree}
\title{Phylogenetic tree for bears.}
\format{A \code{phylo} object with 13 bear species. Branch lengths are scaled in millions of years. The tip labels match the names in \code{\link{bear_probs}}, an included vector of extinction probabilities.}
\usage{
bear_tree
}
\description{
Example dated phylogenetic tree for late Quaternary bear species.
}
\keyword{datasets}
