% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallorn.R
\docType{package}
\name{mallorn-package}
\alias{mallorn}
\alias{mallorn-package}
\title{Compute Expected Evolutionary Distinctiveness}
\description{
This is simple package to calculate expected Phylogenetic Diversity (expected PD) and expected Evolutionary Distinctiveness (expected ED) for each tip of a phylogenetic tree.
}
\details{
Currently, \pkg{mallorn} has two functions: \code{\link{ePD}} and \code{\link{eED}}. \code{\link{ePD}} calculates expected PD, the expected amount of evolutionary history represented by a community. \code{\link{eED}} calculates expected ED, the expected amount of unique evolutionary history belonging to each taxon.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/ursus-americanus/mallorn}
  \item Report bugs at \url{http://github.com/ursus-americanus/mallorn/issues}
}

}
