% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallorn.R
\docType{package}
\name{mallorn-package}
\alias{mallorn}
\alias{mallorn-package}
\title{Compute Expected Evolutionary Distinctiveness}
\description{
This is simple package to calculate Expected Evolutionary Distinctiveness (EED) for each tip of a phylogenetic tree.
}
\details{
Currently, \pkg{mallorn} has only one function \code{\link{eed}}, which is used to calculate Expected Evolutionary Distinctiveness, the expected amount of unique evolutionary history belonging to each taxa.
}
