% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallorn.R
\docType{package}
\name{mallorn-package}
\alias{mallorn}
\alias{mallorn-package}
\title{Compute Expected Evolutionary Distinctiveness}
\description{
This is simple package to calculate Expected Evolutionary Distinctiveness (expected ED) for each tip of a phylogenetic tree.
}
\details{
Currently, \pkg{mallorn} has only one function \code{\link{expected.ed}}, which is used to calculate Expected Evolutionary Distinctiveness, the expected amount of unique evolutionary history belonging to each taxa.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/ursus-americanus/mallorn}
  \item Report bugs at \url{http://github.com/ursus-americanus/mallorn/issues}
}

}
